Change IDataSource.Count from int to long
- IDataSource.Count: int -> long
- IDataSource.CountAsync: Task<int> -> Task<long>
- Updated DataSourceBase, all sync/async item/batch base classes
- Updated DataSourceDecorator, FutureListDataSource, ProjectedDataSource
- Updated Communications data sources (Protocol, WebSockets)
- Updated TestDataSource in WinFormsTester
- Supports data sources with more than Int32.MaxValue items

+++ src/Sphere10.Framework.Communications/DataSource/ClientWebSocketsDataSource.cs
@@ -15,7 +15,7 @@ using Newtonsoft.Json;
 namespace Sphere10.Framework.Communications;
 
 public class ClientWebSocketsDataSource<TItem> : ProtocolChannelDataSource<TItem> {
-	public override Task<int> CountAsync => throw new NotImplementedException();
+	public override Task<long> CountAsync => throw new NotImplementedException();
 
 	public InitializeDelegate InitializeItem { get; init; }
 	public UpdateDelegate UpdateItem { get; set; }
