+++ src/Sphere10.Framework/DataSource/AsyncBatchDataSourceBase.cs
@@ -89,7 +89,7 @@ public abstract class AsyncBatchDataSourceBase<TItem> : DataSourceBase<TItem> {
  public override Result ValidateRange(IEnumerable<(TItem entity, CrudAction action)> actions)
 		=> ValidateRangeAsync(actions).ResultSafe();
 
-	public override int Count => CountAsync.ResultSafe();
+	public override long Count => CountAsync.ResultSafe();
 
 	public override DataSourceCapabilities Capabilities => CapabilitiesAsync.ResultSafe();
 }
